+++ server/handler/secret.go
@@ -35,28 +35,13 @@ func CreateSecretHandler(depot *mysql.MySQLDepot) http.HandlerFunc {
 			return
 		}
 		if client.Status == "INACTIVE" {
-			duration, err := time.ParseDuration(secret.Available_Period)
-			if err != nil {
-				http.Error(w, err.Error(), http.StatusInternalServerError)
-				return
-			}
 			err = depot.UpdateStatusClient(secret.Target, "ISSUABLE")
 			if err != nil {
 				http.Error(w, err.Error(), http.StatusInternalServerError)
 				return
 			}
 			secret.Type = "ACTIVATE"
-			time.AfterFunc(duration, func() {
-				if err := depot.DeleteSecret(secret.Target); err == nil {
-					depot.UpdateStatusClient(secret.Target, "INACTIVE")
-				}
-			})
 		} else if client.Status == "ISSUED" {
-			duration, err := time.ParseDuration(secret.Available_Period)
-			if err != nil {
-				http.Error(w, err.Error(), http.StatusInternalServerError)
-				return
-			}
 			if _, err := time.ParseDuration(secret.Pending_Period); err != nil {
 				http.Error(w, err.Error(), http.StatusInternalServerError)
 				return
@@ -67,11 +52,6 @@ func CreateSecretHandler(depot *mysql.MySQLDepot) http.HandlerFunc {
 				return
 			}
 			secret.Type = "UPDATE"
-			time.AfterFunc(duration, func() {
-				if err := depot.DeleteSecret(secret.Target); err == nil {
-					depot.UpdateStatusClient(secret.Target, "ISSUED")
-				}
-			})
 		} else {
 			http.Error(w, "Client is not in INACTIVE or ISSUED state", http.StatusInternalServerError)
 			return